Check sqeletor last-updated annotation (if present) to see if secret is updated
If sqeletor manages the secret, the correlationId will be copied from the sqlUser, which is only created on the first deploy of the application. It will therefore not accurately reflect if the secret has been updated.
If sqeletor doesn't manage the secret, naiserator does, and the correlationId is presumed to be updated as expected.

Co-authored-by: Trong Nguyen <trong.huu.nguyen@nav.no>