Scale down application if lag is reasonable, promote when lag is zero
Split the replication lag check in two parts.

The first part runs while the application is still running, and will accept a reasonable lag (16MB) and allow the promote to continue.

The second runs after the application has been scaled to 0, and will only allow promotion to proceed when a lag of 0 bytes is confirmed.